pyproject.toml
[tool.poetry]
name = "asteroid-diameter-prediction"
version = "0.3.0"
description = "Predicting asteroid diameters from the Open Asteroid Dataset https://www.kaggle.com/datasets/basu369victor/prediction-of-asteroid-diameter. Toy repo to learn MLOps best practices."
authors = ["SanchitNardekar <sanchit@motorway.co.uk>"]
readme = "README.md"
repository = "https://github.com/SanchitNardekar/asteroid-diameter-prediction"
homepage = "https://github.com/SanchitNardekar/asteroid-diameter-prediction"

[tool.poetry.dependencies]
python = "3.10.*"
click = "^8.1.7"
pandas = "^2.1.2"
numpy = "^1.26.1"
scikit-learn = "^1.3.2"
google-api-python-client = "^2.105.0"
dvc = "^3.28.0"
dvc-gdrive = "^2.20.0"
loguru = "^0.7.2"
statsmodels = "^0.14.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
make-dataset = "asteroid_diameter_prediction.train.make_dataset:main"
enrich-features = "asteroid_diameter_prediction.train.enrich_features:main"
train-model = "asteroid_diameter_prediction.train.train_model:main"
